(GH-11) SemanticVersion - Add package release version
For proper package release versioning, which allows for a NVR format
(Name Version Release) as part of the packaging, add
PackageReleaseVersion to SemanticVersion. This allows fixing a package
without having to up the version of the package version in a way that
could be confusing for others.

Previously the way to handle package fix version notation was done in a
way that supported passing YYYYMMDD in the fourth version segment.
However that has proved to be a hack at best as the 4th segment could
be used for the version of the package or a versioning scheme from a
package could technically use the fourth segment with a date notation.
While it is rare, it has been seen before on non-Windows platforms and
could be seen in Windows packaging at some point.

Provide the ability to create package versions with `_#` (e.g.
`1.0.0_1`) to signify a revision to the package itself without changing
the version of the package itself.