(GH-10) ensure source uris bypass proxy
NuGet has some interesting ways of working with information and
deciding when to instantiate HttpClient. There is no easy way to ensure
that bypassing a proxy is passed around everywhere, but it is possible
to provide a cache to query whether urls are being bypassed when
creating an HttpClient.

Also ensure that where it can be set, BypassProxy is set appropriately
and passed on with new HttpClient implementations.