(GH-14) Provide override to set IgnoreFailingRepositories
Previously, when passing in repositories to aggregate repo, it would
convert all repos to aggregate repository and they would typically not
have IgnoreFailingRepositories set to true. In turn this would ensure
that passing repositories in would never allow setting
IgnoreFailingRepositories to true and it would be ignored.

When the value is passed, be sure to set it for all aggregate
repositories.