(#2258) Pack all files if no files element in nuspec
When running a choco pack on Linux, if there is not a <files> element
in the nuspec file, then no files (apart from the .nuspec) file are
packed, but we would expect that all files/folder at the location of
the .nuspec file would be included. This wasn't happening due to the
fact that the wrong path separator is being used.