(#23) Prevent creation of empty NuGet.Config file
As Chocolatey, we never want to create the NuGet.Config file, as this is
not used by Chocolatey at all. As such, rather than GetOrCreateDocument,
if the filepath doesn't exist, simply return the default empty
XDocument, which is exactly what would have been returned after creating
the file.

The existing coding style has been adhered to here, and as little
changes as possible have been made.