(#25) Add basic nuspec file for generating package
This is a very basic nuspec file that has just the bare minimum to
generate a valid NuGet package for the Chocolatey-NuGet.Core assembly.
A readme file has been created with the steps that need to be followed
to generate the assembly and create the package, ready for consumption
in the chocolatey/choco repository.