(#33) Add DotNetZip to use for package extraction
This commit add DotNetZip to use for extracting nupkgs.
The System.IO.Packaging code is not able to extract files within
nupkgs that contain special characters like spaces or at symbols.
This causes those files to not be extracted from nupkgs when this
library is used by consumers like Chocolatey CLI. This will cause
prolems when packages packed with future versions of Chocolatey CLI
that use NuGet SDK assemblies as files with special characters are not
encoded due to changes in how NuGet operates, and therefore those
packages would be incompatible with older versions of Chocolatey CLI.
It also would effect the Community Repository services in a similar
way, as they also use this library.

This commit then switches the ZipPackage and OptimizedZipPackage
classes to use DotNetZip as the primary way to extract nupkgs. This
allows for the extraction of files with filenames that have special
characters. If DotNetZip is unable to read a passed in package, it
will fall back to the system Packaging code.

The build.cmd is updated to restore the DotNetZip package because
msbuild was not happy about restoreing it. The strongname script was
also updated to include the DotNetZip assembly in the ilmerged final
assembly so a specific reference to DotNetZip is not required in
projects that consume this library.