(#35) Make operations ordered by dependencies
To ensure that all operations are attempted in the correct order, we
need to sort the PackageOperations based on the dependencies that are
in play.  This will ensure that if Package B depends on Package A, then
Package A should be acted on first, before anything is done with
Package B.  This will also be true for all operations within the
complete dependency tree.

This commit makes use of a new ComparePackageOperations method, which
is passed into the List<T>.Sort method, to ensure that the operations
are ordered based package names, and then a new method called
GetOperationsInDependencyOrder is used to ensure that the install and
uninstall operations are sorted in the correct order.